requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4

# Scheduling
apscheduler==3.10.4

# HTTP
requests==2.31.0
paho-mqtt==1.6.1

# Config / env
python-dotenv==1.0.0
pyyaml==6.0.1
pytz==2024.1

# Google Drive
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0

# Pushover (via requests, no extra lib needed)
# pysolarmanv5 intentionally excluded — SN overflows uint32
